requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.22
qdrant-client==1.15.1
sentence-transformers==3.0.1
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pymupdf==1.24.10
wikipedia==1.4.0